Prepopulate the network dictionary.
One of the tests was failing because the network dictionary didn't have entries for each agent
type. It makes sense to prepopulate it with empty lists for the moment, eventually this will move
to shared state in a protocol factory, so it's not too important that it's a bit hacky now.