Try and fix annoying hang in dbp.test.test_dbp.TestDBP.
dbp.test.test_dbp.TestDBP.test_wait_on_txs* all hang due to self.assertEqual raising something
that then gets swallowed either by the reactor or the test harness. This is a pain, and needs
fixing at some point but patch it up enough to work for the moment and move on.