Add a regression test for Learner.
Learner is very indiscriminate in what it accepts, add a test to ensure that it distinguishes
between proposals with the same proposal number but different values. E.g., if there are 5
acceptors, and a learner receives the following messages:
  Accept(Proposal(1, 2))
  Accept(Proposal(1, 2))
  Accept(Proposal(1, 3))
then it will accept Proposal(1, 3), even though this does not have a majority vote.